Final Code Commit
FAQ ordered list completed, along with the back to top link and  links inside of the web page linked to each other. Email for organization is linked to open in default email application and the word "accommodations"  is linked to the accommodations question on the FAQ page. Comments filled in and webpages complete.